breakup misc
