+++ src/aspire/optimization/__init__.py
@@ -1,0 +1,1 @@
+from .conj_grad import conj_grad, fill_struct
