+++ src/aspire/utils/__init__.py
@@ -1,66 +1,2 @@
-import os.path
-import subprocess
-
+from .misc import ensure, get_full_version, powerset, sha256sum
 from .types import complex_type, real_type, utest_tolerance
-
-
-def ensure(cond, error_message=None):
-    """
-    assert statements in Python are sometimes optimized away by the compiler, and are for internal testing purposes.
-    For user-facing assertions, we use this simple wrapper to ensure conditions are met at relevant parts of the code.
-
-    :param cond: Condition to be ensured
-    :param error_message: An optional error message if condition is not met
-    :return: If condition is met, returns nothing, otherwise raises AssertionError
-    """
-    if not cond:
-        raise AssertionError(error_message)
-
-
-def get_full_version():
-    """
-    Get as much version information as we can, including git info (if applicable)
-    This method should never raise exceptions!
-
-    :return: A version number in the form:
-        <maj>.<min>.<bld>
-            If we're running as a package distributed through setuptools
-        <maj>.<min>.<bld>.<rev>
-            If we're running as a 'regular' python source folder, possibly locally modified
-
-            <rev> is one of:
-                'src': The package is running as a source folder
-                <git_tag> or <git_rev> or <git_rev>-dirty: A git tag or commit revision, possibly followed by a suffix
-                    '-dirty' if source is modified locally
-                'x':   The revision cannot be determined
-
-    """
-    import aspire
-
-    full_version = aspire.__version__
-    rev = None
-    try:
-        path = aspire.__path__[0]
-        if os.path.isdir(path):
-            # We have a package folder where we can get git information
-            try:
-                rev = (
-                    subprocess.check_output(
-                        ["git", "describe", "--tags", "--always", "--dirty"],
-                        stderr=subprocess.STDOUT,
-                        cwd=path,
-                    )
-                    .decode("utf-8")
-                    .strip()
-                )
-            except (FileNotFoundError, subprocess.CalledProcessError):
-                # no git or not a git repo? assume 'src'
-                rev = "src"
-    except:  # nopep8  # noqa: E722
-        # Something unexpected happened - rev number defaults to 'x'
-        rev = "x"
-
-    if rev is not None:
-        full_version += f".{rev}"
-
-    return full_version
