+++ src/aspire/volume/__init__.py
@@ -1,6 +1,7 @@
 import logging
 
 import numpy as np
+from numpy.linalg import qr
 
 import aspire.image
 from aspire.nufft import nufft
@@ -8,8 +9,38 @@ from aspire.utils import ensure
 from aspire.utils.coor_trans import grid_2d
 from aspire.utils.fft import centered_ifft2
 from aspire.utils.matlab_compat import m_reshape
+from aspire.utils.matrix import mat_to_vec, vec_to_mat
 
 logger = logging.getLogger(__name__)
+
+
+def qr_vols_forward(sim, s, n, vols, k):
+    """
+    TODO: Write docstring
+    TODO: Find a better place for this!
+    :param sim:
+    :param s:
+    :param n:
+    :param vols:
+    :param k:
+    :return:
+    """
+    ims = np.zeros((k, n, sim.L, sim.L), dtype=vols.dtype)
+    for ell in range(k):
+        ims[ell] = sim.vol_forward(Volume(vols[ell]), s, n).asnumpy()
+
+    ims = np.swapaxes(ims, 1, 3)
+    ims = np.swapaxes(ims, 0, 2)
+
+    Q_vecs = np.zeros((sim.L ** 2, k, n), dtype=vols.dtype)
+    Rs = np.zeros((k, k, n), dtype=vols.dtype)
+
+    im_vecs = mat_to_vec(ims)
+    for i in range(n):
+        Q_vecs[:, :, i], Rs[:, :, i] = qr(im_vecs[:, :, i])
+    Qs = vec_to_mat(Q_vecs)
+
+    return Qs, Rs
 
 
 class Volume:
