+++ tutorials/examples/cov3d_simulation.py
@@ -9,12 +9,12 @@ from scipy.cluster.vq import kmeans2
 
 from aspire.basis import FBBasis3D
 from aspire.covariance import CovarianceEstimator
+from aspire.denoising import src_wiener_coords
 from aspire.noise import WhiteNoiseEstimator
 from aspire.operators import RadialCTFFilter
 from aspire.reconstruction import MeanEstimator
 from aspire.source.simulation import Simulation
 from aspire.utils.matrix import eigs
-from aspire.utils.misc import src_wiener_coords
 from aspire.utils.random import Random
 from aspire.volume import Volume
